moving socketcontext down the component tree
