+++ packages/client/src/components/molecules/LogoStatusMolecules/StatusComponent.tsx
@@ -1,3 +1,4 @@
+import { useEffect, useState } from "react";
 import { twMerge } from "tailwind-merge";
 
 import AWSIcon from "@/components/atoms/AWSIcon";
@@ -6,18 +7,65 @@ import LatencyDotsIcon from "@/components/atoms/LatencyDotsIcon";
 import UserComputerIcon from "@/components/atoms/UserComputerIcon";
 import { CONNECTIONTYPES, useAppState } from "@/contexts/AppStateContext";
 import { usePacket } from "@/contexts/PacketContext";
+import { socketIO } from "@/contexts/SocketIOConnection";
 
 function StatusComponent() {
   const { currentAppState } = useAppState();
   const { currentPacket } = usePacket();
-  const userConnection = currentAppState.socketConnected;
-  // TODO: change carConnection from socketIO.connected to carConnection.connected
-  const carConnection = currentAppState.socketConnected;
+
+  const [latencies, setLatencies] = useState({
+    userLatency: null || 0,
+    carLatency: null || 0,
+    socketConnected: false, // TODO: differentiate between user and car connection
+  });
   const colorTheme = currentAppState.darkMode ? "#FFFFFF" : "#000000";
   // Maybe server should have a reference to the last packet received from the vehicle.
-  const packetTime = currentAppState.socketConnected
+  const packetTime = latencies.socketConnected
     ? new Date(currentPacket.TimeStamp).toLocaleString()
     : "DISCONNECTED";
+
+  const onCarLatency = (latency: number) => {
+    setLatencies((prev) => ({ ...prev, carLatency: latency }));
+  };
+
+  useEffect(() => {
+    // Connect to the socket
+    socketIO.connect();
+
+    socketIO.on("connect", () => {
+      setLatencies((prev) => ({
+        ...prev,
+        socketConnected: true,
+      }));
+    });
+
+    socketIO.on("disconnect", () => {
+      setLatencies((prev) => ({
+        ...prev,
+        socketConnected: false,
+      }));
+    });
+
+    // Ping the server every second to measure user latency
+    const id = setInterval(() => {
+      const start = Date.now();
+
+      socketIO.emit("ping", () => {
+        const duration = Date.now() - start;
+        setLatencies((prev) => ({ ...prev, userLatency: duration }));
+      });
+    }, 1000);
+    socketIO.on("carLatency", onCarLatency);
+    return () => {
+      clearInterval(id);
+      socketIO.off("connect");
+      socketIO.off("disconnect");
+      socketIO.off("carLatency", onCarLatency);
+      socketIO.disconnect();
+    };
+    // Register event listeners
+  }, [setLatencies]);
+
   return (
     <div className="grid">
       <div>
@@ -27,16 +75,16 @@ function StatusComponent() {
             color={colorTheme}
             width="15px"
             height="20px"
-            latency={currentAppState.userLatency}
-            isConnected={userConnection}
+            latency={latencies.userLatency}
+            isConnected={latencies.socketConnected}
           />
           <AWSIcon color={colorTheme} width="25px" height="25px" />
           <LatencyDotsIcon
             color={colorTheme}
             width="15px"
             height="20px"
-            latency={currentAppState.carLatency}
-            isConnected={carConnection}
+            latency={latencies.carLatency}
+            isConnected={latencies.socketConnected}
           />
           <CarIcon color={colorTheme} width="25px" height="25px" />
         </div>
