+++ packages/client/src/pages/_app.tsx
@@ -3,7 +3,6 @@ import type { AppProps } from "next/app";
 import { LoadingWrapper } from "@/components/global/LoadingWrapper";
 import { AppStateContextProvider } from "@/contexts/AppStateContext";
 import { PacketContextProvider } from "@/contexts/PacketContext";
-import { SocketContextProvider } from "@/contexts/SocketContext";
 import "@/styles/globals.css";
 
 export default function App({ Component, pageProps }: AppProps) {
